requirements/common.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
marshmallow
python-dotenv
requests